Moved beans to GatewayServerMvcAutoConfiguration
